requirements.txt
# Django Dependencies
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
sqlparse==0.5.3
django-axes==6.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
django-debug-toolbar==4.3.0

# FastAPI Dependencies
fastapi[standard]==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0

# Database Dependencies
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
databases[postgresql]==0.9.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# File Upload & Storage
aiofiles==24.1.0
pillow==10.4.0

# Testing Dependencies
pytest>=8.2.0
pytest-django>=4.8.0
pytest-asyncio>=0.24.0
httpx==0.27.2

# Utility Dependencies
email-validator==2.2.0
redis==5.1.1
